% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_err.R
\name{compute.rec.err}
\alias{compute.rec.err}
\title{REC_ERR}
\usage{
compute.rec.err(data.mat, n.components)
}
\arguments{
\item{data.mat}{An N x M matrix, containing N rows (observations) and M
columns (data features)}

\item{n.components}{An integer representing the number of principal
components to calculate and measure RE for. Must not exceed M.}
}
\value{
a two-column dataframe containing  \code{n.principal.components} and \code{rec.error} data
}
\description{
Calculates reconstruction error (RE) of PCA for a feature matrix.
}
\examples{
# Prep data
unlabeled.iris.data <- iris[1:4]
num.vars <- ncol(unlabeled.iris.data)

# Compute RE on the unlabeled iris data
rec.err.df <- compute.rec.err(unlabeled.iris.data, num.vars)

# Plot RE over model complexity
plot(rec.err.df$n.principal.components, rec.err.df$rec.error,
     xlab="# PCs", ylab="Reconstruction Err")
}
