% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_prog_cpd.R
\name{DYNPROG}
\alias{DYNPROG}
\title{DYNPROG}
\usage{
DYNPROG(data_vec, kmax)
}
\arguments{
\item{data_vec}{A numeric vector of signal or time-series data with indices of length T}

\item{kmax}{An integer value: the maximum number of segments to be computed}
}
\value{
a matrix of cost values with K rows and T columns
}
\description{
Performs optimal changepoint detection, using the dynamic programming-based
\code{Opt} algorithm described in Truong et al. (2020). Full text available at:
\url{https://arxiv.org/pdf/1801.00718.pdf}
}
\examples{
# TODO: make some examples

}
