% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{PCA}
\alias{PCA}
\title{PCA}
\usage{
PCA(data.matrix)
}
\arguments{
\item{data.matrix}{An N x M matrix, containing N rows (observations) and M
columns (data features)}
}
\value{
A \strong{list} containing the following results:

\item{\code{center}}{a length-M named vector representing the column means of the
input data after centering}
\item{\code{rotation}}{an MxM dataframe of PCA loadings}
\item{\code{x}}{an NxM dataframe of Principal Component vectors}
}
\description{
Performs Principal Components Analysis (PCA) of a given feature matrix, using
Singular Value Decomposition (SVD). Principal components are ordered by
weight, allowing the projection of data onto some user-selected number of
principal components, where PC1 is the line that describes the most
variance in the data, PC2 the second-most variance, etc. This property
makes PCA a valuable tool for dimensionality reduction and
plotting/interrogation of high-dimensional data.
}
\details{
By default, this implementation performs centering but not scaling of data.
If your data requires scaling, see \code{?scale}.
}
\examples{
# Compute principal coordinates of the (unlabeled) iris data
iris.fit <- PCA(iris[1:4])
iris.fit$center
head(iris.fit$rotation)
head(iris.fit$x)
iris.pcs <- as.data.frame(iris.fit[["x"]])

# re-apply labels...
iris.pcs[, "species"] <- iris[5]

# ... then plot a 2D projection of the data

}
